Feature/api updates agentic googlepay compliance digitalcards (#537)
* Add Agentic Commerce endpoint support

New AgenticCommerceClient with delegated payment operations,
including request/response models, entities, and unit + integration
+ serialization tests.

* Add Compliance Requests endpoint support

New ComplianceRequestsClient with get-details and respond operations,
including request/response models and unit + integration + serialization tests.

* Add Google Pay endpoint support

New GooglePayClient with enrollment, domain registration and domain
listing operations, including entities, request/response models,
and unit + integration + serialization tests.

* Add Digital Cards support to Issuing module

New partial IssuingClient extension for digital card operations
(get digital card), including response models, device/requestor
entities, and unit + integration + serialization tests.

* Update Payment Setups and Flow models

Add new properties to PaymentSetups entities (Customer, MerchantAccount,
Industry, Order, Settings, PaymentMethods, etc.), add Billing and
AccommodationData models, add Paypal payment method, update
PaymentSessionSubmitRequest, and add serialization tests.

* Update Authentication client with session headers support

Add SessionHeaders model, extend IAuthenticationClient and
AuthenticationClient with new operations, and update unit tests.

* Wire up new clients and update common infrastructure

Register AgenticCommerce, ComplianceRequests, GooglePay, and
DigitalCards clients in CheckoutApi/ICheckoutApi. Update ApiClient
and IApiClient with new HTTP method overloads. Extract IHeaders
interface, add SchemeLocalType values, update SchemeMetadata,
add nuget.config, update test project references, and fix minor
test adjustments.

* Fix date fields to use DateTime? for consistency

Align MerchantAccount (RegistrationDate, LastModified,
FirstTransactionDate, LastTransactionDate) and OrderSubMerchant
(RegistrationDate) with the existing SDK convention (DateTime?
instead of string), matching PaymentContextsCustomerSummary.
Update serialization test assertions accordingly.

* Apply PR review feedback: enum formatting and CountryCode typing

Place [EnumMember] attributes on their own line above enum values
per project convention. Use CountryCode enum instead of string for
ISO 3166-1 alpha-2 country fields in billing/accommodation addresses.

* Refactor header resolution and enhance enum handling

Introduce a new method, ResolveHeaderValue, to properly handle enum values using [EnumMember] attributes for consistent casing in HTTP headers. Update unit tests to validate the resolution of all channel types in SessionHeaders. Remove unnecessary JsonConverter attributes from enums in PaymentSetups entities and replace string country codes with CountryCode enum for better type safety.

* Refactor enum formatting for consistency

Align [EnumMember] attributes with project conventions by placing them on their own lines above enum values in DelegatedCardFundingType, DelegatedCardNumberType, and SchemeLocalType. This improves readability and maintains consistency across the codebase.

+++ src/CheckoutSdk/HandlePaymentsAndPayouts/PaymentSetups/Entities/Customer/MerchantAccount.cs
@@ -6,42 +6,50 @@ namespace Checkout.Payments.Setups.Entities
     {
         /// <summary>
         /// The merchant's unique identifier for the customer's account
+        /// [Optional]
         /// </summary>
         public string Id { get; set; }
 
         /// <summary>
-        /// The date the customer registered their account with the merchant
+        /// The date the customer registered their account with the merchant (yyyy-MM-dd).
+        /// [Optional]
         /// </summary>
         public DateTime? RegistrationDate { get; set; }
 
         /// <summary>
-        /// The date the customer's account with the merchant was last modified
+        /// The date the customer's account with the merchant was last modified (yyyy-MM-dd).
+        /// [Optional]
         /// </summary>
         public DateTime? LastModified { get; set; }
 
         /// <summary>
         /// Specifies if the customer is a returning customer
+        /// [Optional]
         /// </summary>
         public bool? ReturningCustomer { get; set; }
 
         /// <summary>
-        /// The date of the customer's first transaction
+        /// The date of the customer's first transaction (yyyy-MM-dd).
+        /// [Optional]
         /// </summary>
         public DateTime? FirstTransactionDate { get; set; }
 
         /// <summary>
-        /// The date of the customer's most recent transaction
+        /// The date of the customer's most recent transaction (yyyy-MM-dd).
+        /// [Optional]
         /// </summary>
         public DateTime? LastTransactionDate { get; set; }
 
         /// <summary>
         /// The total number of orders made by the customer
+        /// [Optional]
         /// </summary>
         public int? TotalOrderCount { get; set; }
 
         /// <summary>
         /// The payment amount of the customer's most recent transaction
+        /// [Optional]
         /// </summary>
-        public long? LastPaymentAmount { get; set; }
+        public decimal? LastPaymentAmount { get; set; }
     }
 }
